[ckui] Update main UI look, add README (#274)
* Made the main UI looking better by default by hiding additional information

* Added example data

Example data will be used when no graph_dump.json and / or prometheus_metrics.json is found in the data/ folder.
It's also possible to force the use of example data (for data-sensitive demos) by setting 'use_example_data = true' in the 'singleton_global'.

Other minor changes with this commit:
- The background image was replaced with a radial gradient shader for a smoother look.
- Added the temporary Cloudkeeper logo as a project icon.
- Fixed a couple of warnings.
- Fixed the missing fonts in some resources.
- Removed the camera movement in SFDP calculation.
- Replaced noise texture and it's use.
- Removed a lot of unused graphics.

* Added a useful readme