Blast Radius UI and cleanup (#306)
* Refactored graph view into separate view and graph class.

Connections and Nodes are now part of a CloudGraph class that manages all the neccessary things to display a graph.
Navigation of the main graph was moved from the main UI to a UIGraph (Which is mostly used to display the big main graph and control the camera (and spaceship).
This makes it much easier to generate new sub-graphs and have the navigation separated from the actual graph view and generation logic.

* Cleaned up the UI and made the bottom (UICommandLine) smaller

* Fixed graph refactor bugs

* Added Blastradius logic

* Renamed "connections" to "edges".

* Basic implementation of Blast Radius

* Fixed shallow copy problems with graph_data.

This led to a lot of problem when using more than one graph display.
The best way seems to be to make a refactor of the graph_view and separate generation and layout logic into specific subclasses (main view, blastradius view, list view...)

* Finished first version of the "Blast Radius" ui.

* Fixed bug when clicking on "clear blast radius" node in the blast radius ui